test: keep cancellation regression out of unasync inputs
