+++ tests/_async/test_connection_pool.py
@@ -1,15 +1,12 @@
 import logging
 import typing
 
-import anyio
 import hpack
 import hyperframe.frame
 import pytest
 import trio as concurrency
 
 import httpcore
-from httpcore._async import connection_pool
-from httpcore._async.interfaces import AsyncConnectionInterface
 
 
 @pytest.mark.anyio
@@ -452,75 +449,6 @@ async def test_connection_pool_with_connect_exception():
         "connection.connect_tcp.started",
         "connection.connect_tcp.failed",
     ]
-
-
-@pytest.mark.anyio
-async def test_connection_pool_removes_connection_after_request_cancellation(
-    monkeypatch: pytest.MonkeyPatch,
-) -> None:
-    request_assigned = anyio.Event()
-
-    class WaitingPoolRequest(connection_pool.AsyncPoolRequest):
-        async def wait_for_connection(
-            self, timeout: float | None = None
-        ) -> AsyncConnectionInterface:
-            await super().wait_for_connection(timeout)
-            request_assigned.set()
-            await anyio.sleep_forever()
-            raise AssertionError("Pool request must be cancelled")
-
-    monkeypatch.setattr(connection_pool, "AsyncPoolRequest", WaitingPoolRequest)
-
-    class ConnectingConnection(AsyncConnectionInterface):
-        def __init__(self) -> None:
-            self.closed = False
-
-        async def handle_async_request(
-            self, request: httpcore.Request
-        ) -> httpcore.Response:
-            raise AssertionError("Cancelled request must not reach the connection")
-
-        async def aclose(self) -> None:
-            self.closed = True
-
-        def can_handle_request(self, origin: httpcore.Origin) -> bool:
-            return True
-
-        def is_available(self) -> bool:
-            return False
-
-        def has_expired(self) -> bool:
-            return False
-
-        def is_idle(self) -> bool:
-            return False
-
-        def is_closed(self) -> bool:
-            return False
-
-        def info(self) -> str:
-            return "CONNECTING"
-
-    class TestPool(httpcore.AsyncConnectionPool):
-        def create_connection(
-            self, origin: httpcore.Origin
-        ) -> AsyncConnectionInterface:
-            return connection
-
-    connection = ConnectingConnection()
-    async with TestPool(max_connections=1) as pool:
-
-        async def request() -> None:
-            with pytest.raises(TimeoutError):
-                with anyio.fail_after(0.01):
-                    await pool.request("GET", "https://example.com/")
-
-        async with anyio.create_task_group() as task_group:
-            task_group.start_soon(request)
-            await request_assigned.wait()
-
-        assert pool.connections == []
-        assert connection.closed
 
 
 @pytest.mark.anyio
